requirements.txt
Django==4.2.18
ruff==0.9.2
django-allauth[socialaccount]==65.3.1
django-silk==5.3.1
mysqlclient==2.2.7
